Update README.md
The  Unsplash Clone is a web application built with **Nuxt.js** that allows users to search for high-quality images from the **Unsplash API**. Users can view images in a grid layout, click on them to see larger versions in a modal with additional details about the author and location, and experience a smooth loading experience with a shimmer effect. The app features a responsive design and utilizes **Pinia** for state management and **Axios** for API requests.